Continue porting the RTS

+++ rts/motoko-rts/src/blob_iter.rs
@@ -5,9 +5,9 @@ use crate::{
 
 use motoko_rts_macros::ic_mem_fn;
 
-const ITER_BLOB_IDX: u32 = 0;
+const ITER_BLOB_IDX: usize = 0;
 
-const ITER_POS_IDX: u32 = 1;
+const ITER_POS_IDX: usize = 1;
 
 /// Returns iterator for the given blob
 #[ic_mem_fn]
@@ -28,7 +28,7 @@ unsafe fn blob_iter<M: crate::memory::Memory>(mem: &mut M, blob: Value) -> Value
 
 /// Returns whether the iterator is finished
 #[no_mangle]
-unsafe extern "C" fn blob_iter_done(iter: Value) -> u32 {
+unsafe extern "C" fn blob_iter_done(iter: Value) -> usize {
     let iter_array = iter.as_array();
 
     let blob = iter_array.get(ITER_BLOB_IDX);
@@ -39,7 +39,7 @@ unsafe extern "C" fn blob_iter_done(iter: Value) -> u32 {
 
 /// Reads next byte, advances the iterator
 #[no_mangle]
-unsafe fn blob_iter_next(iter: Value) -> u32 {
+unsafe fn blob_iter_next(iter: Value) -> usize {
     let iter_array = iter.as_array();
 
     let blob = iter_array.get(ITER_BLOB_IDX);
