+++ rts/motoko-rts/src/gc.rs
@@ -18,9 +18,9 @@ unsafe fn should_do_gc(max_live: crate::types::Bytes<usize>) -> bool {
     const HEAP_GROWTH_FACTOR: f64 = 1.5;
 
     let heap_limit = core::cmp::min(
-        (usize::from(LAST_HP) * HEAP_GROWTH_FACTOR) as u64,
-        (usize::from(LAST_HP) + max_live.as_usize()) / 2,
+    (LAST_HP as f64 * HEAP_GROWTH_FACTOR) as usize,
+        (LAST_HP + max_live.as_usize()) / 2
     );
 
-    u64::from(HP) >= heap_limit
+    HP >= heap_limit
 }
